Ajout des controles des commandes avec des boutons créés dynamiquement en fonction des commandes du périphérique

+++ app/src/main/java/com/monfort/projetpolyhome/data/DeviceData.kt
@@ -6,7 +6,7 @@ data class DeviceData(
     val availableCommands: List<String>,
 
     // pas dispo pour toutes les commandes
-    val opening: Int? = null,
+    val opening: Float? = null,
     val openingMode: Int? = null,
     val power: Int? = null
 )
